Rename RideConstruction.cpp for MSVC compilation
