+++ src/openrct2-ui/libopenrct2ui.vcxproj
@@ -61,7 +61,7 @@
     <ClInclude Include="interface\Viewport.h" />
     <ClInclude Include="interface\Widget.h" />
     <ClInclude Include="interface\Window.h" />
-    <ClInclude Include="ride\RideConstruction.h" />
+    <ClInclude Include="ride\Construction.h" />
     <ClInclude Include="scripting\CustomImages.h" />
     <ClInclude Include="scripting\CustomListView.h" />
     <ClInclude Include="scripting\CustomMenu.h" />
@@ -120,7 +120,7 @@
     <ClCompile Include="interface\ViewportInteraction.cpp" />
     <ClCompile Include="interface\Widget.cpp" />
     <ClCompile Include="interface\Window.cpp" />
-    <ClCompile Include="ride\RideConstruction.cpp" />
+    <ClCompile Include="ride\Construction.cpp" />
     <ClCompile Include="scripting\CustomImages.cpp" />
     <ClCompile Include="scripting\CustomListView.cpp" />
     <ClCompile Include="scripting\CustomMenu.cpp" />
