+++ src/openrct2-ui/ride/Construction.h
